Move slice query def to depmap_compute
