+++ depmap-compute/depmap_compute/context.py
@@ -6,6 +6,8 @@ from json_logic import jsonLogic, operations  # type: ignore
 from typing import Any, Callable, Union
 from urllib.parse import unquote
 
+from depmap_compute.slice import SliceQuery
+
 
 # Custom JsonLogic operators
 operations.update(
@@ -35,7 +37,7 @@ class ContextEvaluator:
     """
 
     def __init__(
-        self, context: dict, get_slice_data: Callable[[dict[str, str]], dict[str, Any]],
+        self, context: dict, get_slice_data: Callable[[SliceQuery], dict[str, Any]],
     ):
         """
         A `context` dict should have:
@@ -94,7 +96,7 @@ class _JsonLogicVarLookup(dict):
         context_type: str,
         given_id: str,
         cache: dict,
-        get_slice_data: Callable[[str], dict[str, Any]],
+        get_slice_data: Callable[[SliceQuery], dict[str, Any]],
     ):
         self.context_type = context_type
         self.given_id = given_id
@@ -119,6 +121,9 @@ class _JsonLogicVarLookup(dict):
         else:
             cache_key = _encode_slice_query(context_var)
             if cache_key not in self.cache:
+                slice_query_obj = SliceQuery(
+                    **context_var
+                )  # TODO: does this constructor work okay with the enum? - if not, make literal
                 self.cache[cache_key] = self.get_slice_data(context_var)
 
             slice_values = self.cache[cache_key]
