+++ portal-backend/depmap/data_access/__init__.py
@@ -1,6 +1,4 @@
 from .interface import (
-    SliceQuery,
-    SliceIdentifierType,
     # methods that will likely be supported going forward
     get_all_matrix_datasets,
     get_matrix_dataset,
