+++ portal-backend/depmap/data_access/models.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from enum import Enum
 from typing import Optional
 
 
@@ -20,19 +19,3 @@ class MatrixDataset:
     taiga_id: Optional[str]
     units: Optional[str]
     is_continuous: bool
-
-
-# These type definitions will eventually be moved to the shared module
-class SliceIdentifierType(Enum):
-    feature_id = "feature_id"
-    feature_label = "feature_label"
-    sample_id = "sample_id"
-    sample_label = "sample_label"
-    column = "column"
-
-
-@dataclass
-class SliceQuery:
-    dataset_id: str
-    identifier: str
-    indentifier_type: SliceIdentifierType
